% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_RMS.R
\name{calc_RMS}
\alias{calc_RMS}
\title{Calculates various roughness parameters}
\usage{
calc_RMS(data, ...)
}
\arguments{
\item{data}{input data}

\item{...}{currently not used}
}
\value{
console output
}
\description{
Calculates various roughness parameters
}
\examples{
# none

}

