% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Roughness.R
\name{calc_Roughness}
\alias{calc_Roughness}
\title{Calculates various roughness parameters}
\usage{
calc_Roughness(data, method = c("rms"), ...)
}
\arguments{
\item{data}{input data}

\item{method}{which roughness values should be calculated}

\item{...}{currently not used}
}
\value{
console output
}
\description{
Calculates various roughness parameters
}
\examples{
# none
}

