% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Stats.R
\name{calc_Stats}
\alias{calc_Stats}
\title{Calculates basic statistics}
\usage{
calc_Stats(data)
}
\arguments{
\item{data}{input data}
}
\value{
A data frame
}
\description{
Calculates basic statistics
}
\examples{
# none
}

