% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend_Data.R
\name{detrend_Data}
\alias{detrend_Data}
\title{Linear regression to detrend the data}
\usage{
detrend_Data(data, plot = TRUE)
}
\arguments{
\item{data}{input data}

\item{plot}{TRUE or FALSE}
}
\value{
console output
}
\description{
Linear regression to detrend the data
}
\examples{

# none availble yet

}

