% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Data.R
\name{import_Data}
\alias{import_Data}
\title{Function to import .log files}
\usage{
import_Data(file, warnings = FALSE)
}
\arguments{
\item{file}{full path to the file}

\item{warnings}{show warnings produced by readLines (TRUE or FALSE)}
}
\value{
a data.frame
}
\description{
Insert valuable information here
}
\examples{
# none

}

