% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CorrelationLength.R
\name{calc_CorrelationLength}
\alias{calc_CorrelationLength}
\title{Calculate the Correlation length}
\usage{
calc_CorrelationLength(data, plot = TRUE)
}
\arguments{
\item{data}{\code{\link{data.frame}} (\bold{required}): 
A log file imported with \code{\link{import_Data}}.}

\item{plot}{\code{\link{logical}} (optional): \code{TRUE} to plot
the graphs produced by \code{\link{acf}}.}
}
\value{
A plot and a \code{\link{numeric}} value of the correlation length.
}
\description{
Function to calculate the correlation length.
}
\examples{

# Load example data
file <- system.file("extdata/raw.log", package = "roughness")
data <- import_Data(file)

# find and remove outliers
data_cleaned <- find_Outliers(data = data,
                              prefilter = TRUE,
                              method = c("iqr", "sd", "mad"),
                              remove = "mad",
                              window = FALSE,
                              width = 20,
                              step = 10, 
                              hist = TRUE,
                              plot = TRUE)
                              
# de-trend the data set
data_detrend <- detrend_Data(data_cleaned, plot = TRUE)

# calculate RMS
CL <- calc_CorrelationLength(data_detrend, plot = TRUE)
print(CL)

}
