% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CorrelationLength.R
\name{calc_CorrelationLength}
\alias{calc_CorrelationLength}
\title{Calculate the Correlation length}
\usage{
calc_CorrelationLength(data, plot = TRUE)
}
\arguments{
\item{data}{input data}

\item{plot}{TRUE or FALSE}
}
\value{
console output
}
\description{
<placeholder>
}
\details{
<placeholder>
}
\examples{

# none available yet

}

