% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_RMS.R
\name{calc_RMS}
\alias{calc_RMS}
\title{Calculate Root Mean Square Height after Davidson et al. 2003}
\usage{
calc_RMS(data, ...)
}
\arguments{
\item{data}{\code{\link{data.frame}} (\bold{required}): 
A log file imported with \code{\link{import_Data}}.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{data.frame}} containing the RMS value.
}
\description{
This function calculates the Root Mean Square (RMS) Height after Davidson et al. 2003.
}
\examples{
# none

}

