% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_RMS.R
\name{calc_RMS}
\alias{calc_RMS}
\title{Calculate Root Mean Square Height after Davidson et al. 2003}
\usage{
calc_RMS(data, ...)
}
\arguments{
\item{data}{\code{\link{data.frame}} (\bold{required}): 
A log file imported with \code{\link{import_Data}}.}

\item{...}{Currently not used.}
}
\value{
A \code{\link{data.frame}} containing the RMS value.
}
\description{
This function calculates the Root Mean Square (RMS) Height after Davidson et al. 2003.
}
\examples{

# Load example data
file <- system.file("extdata/raw.log", package = "roughness")
data <- import_Data(file)

# find and remove outliers
data_cleaned <- find_Outliers(data = data,
                              prefilter = TRUE,
                              method = c("iqr", "sd", "mad"),
                              remove = "mad",
                              window = FALSE,
                              width = 20,
                              step = 10, 
                              hist = TRUE,
                              plot = TRUE)
                              
# de-trend the data set
data_detrend <- detrend_Data(data_cleaned, plot = TRUE)

# calculate RMS
RMS <- calc_RMS(data_detrend)
print(RMS)

# note that the function returns a data frame
str(RMS)

}

