% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Stats.R
\name{calc_Stats}
\alias{calc_Stats}
\title{Calculate basic statistics}
\usage{
calc_Stats(data)
}
\arguments{
\item{data}{\code{\link{data.frame}} (\bold{required}): 
A log file imported with \code{\link{import_Data}}.}
}
\value{
A \code{\link{data.frame}} with information on the mean, median,
standard deviation, relative standard deviation, variance, range,
minimum and maximum value and slope of a linear regression.
}
\description{
This function calculates several descriptive statistics.
}
\examples{
# none
}

