% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend_Data.R
\name{detrend_Data}
\alias{detrend_Data}
\title{Linear regression to detrend the data}
\usage{
detrend_Data(data, plot = TRUE)
}
\arguments{
\item{data}{\code{\link{data.frame}} (\bold{required}): 
A log file imported with \code{\link{import_Data}}.}

\item{plot}{\code{\link{logical}} (optional): \code{TRUE} to plot
the data set before and after detrending.}
}
\value{
A \code{\link{data.frame}} with detrended input data.
}
\description{
This function performs a linear regression to detrend the data set.
}
\examples{

# Load example data
file <- system.file("extdata/raw.log", package = "roughness")
data <- import_Data(file)

# find and remove outliers
data_cleaned <- find_Outliers(data = data,
                              prefilter = TRUE,
                              method = c("iqr", "sd", "mad"),
                              remove = "mad",
                              window = FALSE,
                              width = 20,
                              step = 10, 
                              hist = TRUE,
                              plot = TRUE)
                              
# de-trend the data set
data_detrend <- detrend_Data(data_cleaned, plot = TRUE)

}

