% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_Outliers.R
\name{find_Outliers}
\alias{find_Outliers}
\title{Function to detect outliers in the roughness data}
\usage{
find_Outliers(data, prefilter = TRUE, method = c("iqr", "sd", "mad"),
  remove = "mad", window = FALSE, width = 20, step = 10, hist = TRUE,
  plot = TRUE, ...)
}
\arguments{
\item{data}{\code{\link{data.frame}} (\bold{required}): 
A log file imported with \code{\link{import_Data}}.}

\item{prefilter}{\code{\link{logical}} (optional): 
\code{TRUE} to remove values that are +- 10 cm of the median}

\item{method}{\code{\link{character}} (optional): 
Vector of methods applied to detect outliers.
Currently implemented: \code{"iqr"} for Interquartile Range (see \code{\link{IQR}}),
 \code{"sd"} for Standard Deviation or \code{"mad"} for
Median Absolut Deviation.}

\item{remove}{\code{\link{character}} (optional):
Returns the original data set with ourliers removed by the
specified method (same options as for argument \code{method}).}

\item{window}{\code{\link{logical}} (optional):
If \code{TRUE} the data set is split in chunks of size specified by \code{width}
before applying any of the detection methods.}

\item{width}{\code{\link{integer}} (optional): Width of moving window.}

\item{step}{\code{\link{integer}} (optional): Step size of moving window.}

\item{hist}{\code{\link{logical}} (optional): Should histograms be plotted?}

\item{plot}{\code{\link{logical}} (optional): Should the results be plotted?}

\item{...}{Currently unused.}
}
\value{
The data set with outliers removed is returned. Optionally, a plot
is created
}
\description{
This function applies three different methods to detect outliers.
}
\examples{

# Load example data
file <- system.file("extdata/raw.log", package = "roughness")
data <- import_Data(file)

# find outliers
data_cleaned <- find_Outliers(data = data,
                              prefilter = TRUE,
                              method = c("iqr", "sd", "mad"),
                              remove = "mad",
                              window = FALSE,
                              width = 20,
                              step = 10, 
                              hist = TRUE,
                              plot = TRUE)

head(data_cleaned)

# find outliers with window technique
data_cleaned2 <- find_Outliers(data = data,
                               prefilter = TRUE,
                               method = "mad",
                               remove = "mad",
                               window = FALSE,
                               width = 30,
                               step = 15, 
                               hist = TRUE,
                               plot = TRUE)
                               
head(data_cleaned2)

}

