% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_Data.R
\name{import_Data}
\alias{import_Data}
\title{Function to import .log files}
\usage{
import_Data(file, warnings = FALSE)
}
\arguments{
\item{file}{\code{\link{character}} (\bold{required}): Full path to a log file.}

\item{warnings}{\code{\link{logical}} (optional): 
\code{TRUE} to show warnings produced by \code{\link{readLines}}.}
}
\value{
Returns a \code{\link{data.frame}}.
}
\description{
This function imports a .log file to R and produces a \code{data.frame} to 
be used for further processing.
}
\note{
The error code \code{65534} is converted to \code{\link{NA}}.
}
\examples{
# none

}

